+++ app/src/components/CreateOrUpdateMeetingForm/CreateOrUpdateMeetingForm.tsx
@@ -428,10 +428,6 @@ const CreateOrUpdateMeetingForm = ({ existingEvent, onForwardButtonClick }: Crea
             values.sharedFolder = false;
             setIsFirstTryToCreateSharedFolder(true);
           },
-          SnackbarProps: {
-            role: 'alert',
-            'aria-label': t('dashboard-meeting-shared-folder-create-error-message'),
-          },
         });
         return false;
       }
@@ -466,10 +462,6 @@ const CreateOrUpdateMeetingForm = ({ existingEvent, onForwardButtonClick }: Crea
           onCancel: () => {
             values.sharedFolder = true;
             setIsFirstTryToDeleteSharedFolder(true);
-          },
-          SnackbarProps: {
-            role: 'alert',
-            'aria-label': t('dashboard-meeting-shared-folder-delete-error-message'),
           },
         });
         return false;