+++ app/src/components/TimerPopover/fragments/CoffeeBreakPopover.tsx
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { styled, Typography, Popover as MuiPopover, Box } from '@mui/material';
+import { styled, Typography, Stack } from '@mui/material';
 import { CoffeeBreakIcon as CoffeeBreakIconDefault } from '@opentalk/common';
 import { TimerStyle } from '@opentalk/common';
 import { useTranslation } from 'react-i18next';
@@ -9,22 +9,7 @@ import { useTranslation } from 'react-i18next';
 import { useAppSelector } from '../../../hooks';
 import { selectTimerActive } from '../../../store/slices/timerSlice';
 import TimerDuration from '../../TimerTab/fragments/TimerDuration';
-
-const Popover = styled(MuiPopover)(({ theme }) => ({
-  pointerEvents: 'none',
-
-  '& .MuiPaper-root': {
-    padding: theme.spacing(2, 4),
-    pointerEvents: 'all',
-  },
-}));
-
-const Content = styled(Box)({
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  gap: '1rem',
-});
+import { Container } from './Container';
 
 const CoffeeBreakIcon = styled(CoffeeBreakIconDefault)(({ theme }) => ({
   height: '4rem',
@@ -32,24 +17,13 @@ const CoffeeBreakIcon = styled(CoffeeBreakIconDefault)(({ theme }) => ({
   fill: theme.palette.secondary.light,
 }));
 
-const CoffeeBreakPopover = ({ anchorEl }: { anchorEl: HTMLElement | null }) => {
+const CoffeeBreakPopover = () => {
   const { t } = useTranslation();
   const timerActive = useAppSelector(selectTimerActive);
 
   return (
-    <Popover
-      role="dialog"
-      aria-labelledby="transition-modal-title"
-      anchorEl={anchorEl}
-      open={Boolean(timerActive)}
-      hideBackdrop
-      anchorOrigin={{
-        vertical: 'top',
-        horizontal: 'right',
-      }}
-      disableEnforceFocus
-    >
-      <Content>
+    <Container open={Boolean(timerActive)} role="dialog" aria-labelledby="transition-modal-title">
+      <Stack alignItems="center" spacing={1}>
         <Typography id="transition-modal-title" variant="h5" component="h2">
           {t('coffee-break-popover-title')}
         </Typography>
@@ -57,8 +31,8 @@ const CoffeeBreakPopover = ({ anchorEl }: { anchorEl: HTMLElement | null }) => {
         <CoffeeBreakIcon />
 
         <TimerDuration alignItems="center" style={TimerStyle.CoffeeBreak} />
-      </Content>
-    </Popover>
+      </Stack>
+    </Container>
   );
 };