+++ app/src/components/TimerPopover/fragments/NormalTimerPopover.tsx
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { Button, Stack, styled, Typography, Popover as MuiPopover } from '@mui/material';
+import { Button, Stack, Typography } from '@mui/material';
 import { TimerStyle } from '@opentalk/common';
 import { useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -16,17 +16,9 @@ import {
 } from '../../../store/slices/timerSlice';
 import { selectOurUuid } from '../../../store/slices/userSlice';
 import TimerDuration from '../../TimerTab/fragments/TimerDuration';
+import { Container } from './Container';
 
-const Popover = styled(MuiPopover)(({ theme }) => ({
-  pointerEvents: 'none',
-
-  '& .MuiPaper-root': {
-    padding: theme.spacing(2, 4),
-    pointerEvents: 'all',
-  },
-}));
-
-const NormalTimerPopover = ({ anchorEl }: { anchorEl: HTMLElement | null }) => {
+const NormalTimerPopover = () => {
   const { t } = useTranslation();
   const dispatch = useAppDispatch();
   const userId = useAppSelector(selectOurUuid);
@@ -43,33 +35,20 @@ const NormalTimerPopover = ({ anchorEl }: { anchorEl: HTMLElement | null }) => {
   }, [timerId, isUserReady]);
 
   return (
-    <Popover
-      anchorEl={anchorEl}
-      open={Boolean(timerId)}
-      hideBackdrop
-      anchorOrigin={{
-        vertical: 'top',
-        horizontal: 'right',
-      }}
-      disableEnforceFocus
-    >
-      {/* 
-        FIXME: this should technically be an alert, but screenreader cannot match the countdown speed,
-        while it is pronouncing the remaining time, the countdown is already over.
-      */}
-      <Stack spacing={2} role="dialog" aria-label={timerTitle || t('timer-popover-title')}>
+    <Container open={Boolean(timerId)} role="dialog" aria-label={timerTitle || t('timer-popover-title')}>
+      <Stack spacing={2}>
         <Typography variant="h1">{t('timer-popover-title')}</Typography>
         {timerTitle && <Typography variant="h2">{timerTitle}</Typography>}
         <TimerDuration style={TimerStyle.Normal} />
         {hasReadyCheckEnabled && (
-          // When timer popover is open we want to focus it so screen reader can tell the content.
+          /* When timer popover is open we want to focus it so screen reader can tell the content. */
           /* eslint-disable jsx-a11y/no-autofocus */
-          <Button onClick={handleDone} color={isUserReady ? 'secondary' : 'primary'} autoFocus>
+          <Button onClick={handleDone} color={isUserReady ? 'secondary' : 'primary'} autoFocus fullWidth>
             {t(`timer-popover-button-${isUserReady ? 'not-' : ''}done`)}
           </Button>
         )}
       </Stack>
-    </Popover>
+    </Container>
   );
 };