+++ app/src/provider/BreakoutRoomProvider/fragments/BreakoutRoomNotification.tsx
@@ -38,6 +38,7 @@ const StyledSnackbarContent = styled(SnackbarContent)(({ theme }) => ({
 
 const BreakoutRoomNotification = React.forwardRef<HTMLDivElement, IJoinNotificationProps>(
   ({ message, actions, iconComponent: Icon, countdown, snackbarKey }, ref) => {
+    const messageId = 'breakout-room-notification-message';
     const [alreadyClicked, setAlreadyClicked] = useState(false);
     const handleOnClick = (onClick: () => void) => {
       !alreadyClicked && onClick();
@@ -54,17 +55,25 @@ const BreakoutRoomNotification = React.forwardRef<HTMLDivElement, IJoinNotificat
 
     const renderActions = () =>
       actions.map(({ text, onClick, variant }, index) => (
-        <Button onClick={() => handleOnClick(onClick)} variant={variant} key={index}>
+        <Button
+          onClick={() => handleOnClick(onClick)}
+          variant={variant}
+          key={index}
+          // Autofocus has been applied to first button of the breakout room notification
+          // as we want to lead the user to the alertdialog without trapping him there.
+          /* eslint-disable jsx-a11y/no-autofocus */
+          autoFocus={index === 0}
+        >
           {text}
         </Button>
       ));
 
     return (
-      <StyledSnackbarContent ref={ref} role="alert">
+      <StyledSnackbarContent ref={ref} role="alertdialog" aria-describedby={messageId}>
         <Stack spacing={1}>
           <Box display="flex" alignItems="center">
             <Icon width="2rem" height="2rem" />
-            <Typography variant="h3" ml={2}>
+            <Typography variant="h3" ml={2} id={messageId}>
               {message}
             </Typography>
           </Box>